# feat: async fetch of components, installers and dependencies