--- a/bottles/backend/managers/installer.py
+++ b/bottles/backend/managers/installer.py
@@ -52,9 +52,9 @@ logging = Logger()
 
 class InstallerManager:
 
-    def __init__(self, manager, offline: bool = False):
+    def __init__(self, manager, offline: bool = False, callback = None):
         self.__manager = manager
-        self.__repo = manager.repository_manager.get_repo("installers", offline)
+        self.__repo = manager.repository_manager.get_repo("installers", offline, callback)
         self.__utils_conn = manager.utils_conn
         self.__component_manager = manager.component_manager
         self.__local_resources = {}