--- a/bottles/backend/managers/repository.py
+++ b/bottles/backend/managers/repository.py
@@ -54,10 +54,10 @@ class RepositoryManager:
         self.__check_locals()
         self.__get_index()
 
-    def get_repo(self, name: str, offline: bool = False):
+    def get_repo(self, name: str, offline: bool = False, callback = None):
         if name in self.__repositories:
             repo = self.__repositories[name]
-            return repo["cls"](repo["url"], repo["index"], offline=offline)
+            return repo["cls"](repo["url"], repo["index"], offline=offline, callback=callback)
 
         logging.error(f"Repository {name} not found")