--- a/bottles/backend/repos/repo.py
+++ b/bottles/backend/repos/repo.py
@@ -15,6 +15,7 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 
+import time
 from bottles.backend.utils import yaml
 import pycurl
 from io import BytesIO
@@ -22,20 +23,27 @@ from http.client import RemoteDisconnected
 
 from bottles.backend.logger import Logger
 
+from bottles.frontend.utils.threading import RunAsync
+
 logging = Logger()
 
 
 class Repo:
     name: str = ""
 
-    def __init__(self, url: str, index: str, offline: bool = False):
+    def __init__(self, url: str, index: str, offline: bool = False, callback = None):
         self.url = url
-        self.catalog = self.__get_catalog(index, offline)
+        self.catalog = None
+
+        def set_catalog(result, error=None):
+            self.catalog = result
+            if callback: callback()
+        RunAsync(self.__get_catalog, callback=set_catalog, index=index, offline=offline)
 
     def __get_catalog(self, index: str, offline: bool = False):
         if index in ["", None] or offline:
             return {}
-
+        time.sleep(3)
         try:
             buffer = BytesIO()