# Description
This is the splitting of PR #2400. This part is only the "backend" part, with only the async fetching. No UI has been updated, so for example preferences menu will be empty until the components are fetched AND the menu is closed and reopened.

## Benchmark
I used `trickle` to limit the bandwidth of Bottles. Example: `trickle -s -d 10 bottles`
Here, the startup time is from the instant that the command is typed, to the instant when the bottles show up. It also counts the time when the window is not showing up.

| Without download limit | Startup time | Updated resources time (after startup) |
|--------------------------|--------------|----------------------------------------|
| Without the PR           | ~2s           | 0s                                     |
| With the PR              | -1s           | 0s                                     |

| Download limit of 50KB/s | Startup time | Updated resources time (after startup) |
|--------------------------|--------------|----------------------------------------|
| Without the PR           | 3s           | 0s                                     |
| With the PR              | 1.5s           | 1s                                     |

| Download limit of 10KB/s | Startup time | Updated resources time (after startup) |
|--------------------------|--------------|----------------------------------------|
| Without the PR           | 14s           | 0s                                     |
| With the PR              | 5s           | 10s                                     |

## Type of change
- [ ] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update