@Kinsteen 2022-12-24
My gift for this Christmas, is an improvement for index fetching! Before, index fetching could take up to 4 seconds (with a good connection) because all of it was sequential. This last commit makes them all parallel, which speeds up a lot this fetching. Combined with the previous async fetch of components, we have some quite performant result